Bug 1377605 Cast -1 to DWORD to correctly match the error return value given by SuspendThread r=aklotz
MozReview-Commit-ID: 5XHgkpQJQKY

UltraBlame original commit: 4cc4de1e589701a48b6e8afe6fae8bb76229ab49

+++ js/src/wasm/WasmSignalHandlers.cpp
@@ -1679,7 +1679,7 @@ js::InterruptRunningJitCode(JSContext* cx)
     
     
     HANDLE thread = (HANDLE)cx->threadNative();
-    if (SuspendThread(thread) != -1) {
+    if (SuspendThread(thread) != (DWORD)-1) {
         CONTEXT context;
         context.ContextFlags = CONTEXT_FULL;
         if (GetThreadContext(thread, &context)) {
